Refactored so that mimimized work is a actual JPA entity.
