+++ orcid-core/src/main/java/org/orcid/core/adapter/JpaJaxbWorkAdapter.java
@@ -21,8 +21,8 @@ import java.util.List;
 
 import org.orcid.jaxb.model.record.summary_rc2.WorkSummary;
 import org.orcid.jaxb.model.record_rc2.Work;
+import org.orcid.persistence.jpa.entities.MinimizedWorkEntity;
 import org.orcid.persistence.jpa.entities.WorkEntity;
-import org.orcid.persistence.jpa.entities.custom.MinimizedWorkEntity;
 
 /**
  *
