+++ orcid-core/src/main/java/org/orcid/core/adapter/impl/JpaJaxbWorkAdapterImpl.java
@@ -24,8 +24,8 @@ import ma.glasnost.orika.MapperFacade;
 import org.orcid.core.adapter.JpaJaxbWorkAdapter;
 import org.orcid.jaxb.model.record.summary_rc2.WorkSummary;
 import org.orcid.jaxb.model.record_rc2.Work;
+import org.orcid.persistence.jpa.entities.MinimizedWorkEntity;
 import org.orcid.persistence.jpa.entities.WorkEntity;
-import org.orcid.persistence.jpa.entities.custom.MinimizedWorkEntity;
 
 /**
  *
