+++ orcid-core/src/main/java/org/orcid/pojo/ajaxForm/WorkForm.java
@@ -38,8 +38,8 @@ import org.orcid.jaxb.model.record_rc2.CitationType;
 import org.orcid.jaxb.model.record_rc2.ExternalID;
 import org.orcid.jaxb.model.record_rc2.Relationship;
 import org.orcid.jaxb.model.record_rc2.Work;
+import org.orcid.persistence.jpa.entities.MinimizedWorkEntity;
 import org.orcid.jaxb.model.record_rc2.ExternalIDs;
-import org.orcid.persistence.jpa.entities.custom.MinimizedWorkEntity;
 import org.orcid.utils.OrcidStringUtils;
 
 public class WorkForm implements ErrorsInterface, Serializable {
