+++ orcid-persistence/src/main/java/org/orcid/persistence/jpa/entities/WorkEntity.java
@@ -26,18 +26,12 @@ import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 
 import org.orcid.jaxb.model.message.CitationType;
 import org.orcid.jaxb.model.message.Iso3166Country;
-import org.orcid.jaxb.model.message.Visibility;
-import org.orcid.jaxb.model.message.WorkType;
 
 /**
  * orcid-entities - Dec 6, 2011 - WorkEntity
@@ -46,102 +40,16 @@ import org.orcid.jaxb.model.message.WorkType;
  */
 @Entity
 @Table(name = "work")
-public class WorkEntity extends BaseEntity<Long> implements Comparable<WorkEntity>, ProfileAware, DisplayIndexInterface, SourceAware {
+public class WorkEntity extends WorkBaseEntity implements Comparable<WorkEntity>, ProfileAware, DisplayIndexInterface, SourceAware {
 
-    private static final long serialVersionUID = -8096348611438944935L;
-    
-    private Long id;
-    private String title;
-    private String translatedTitle;
-    private String subtitle;
-    private String description;
-    private String workUrl;
+    private static final long serialVersionUID = 1L;
+
     private String citation;
-    private String journalTitle;
-    private String languageCode;
-    private String translatedTitleLanguageCode;
     private Iso3166Country iso2Country;
     private CitationType citationType;
-    private WorkType workType;
-    private PublicationDateEntity publicationDate;
     private String contributorsJson;
-    private String externalIdentifiersJson;
     private ProfileEntity profile;
-    private SourceEntity source;
     private Date addedToProfileDate;
-    private Visibility visibility;
-    private Long displayIndex;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO, generator = "work_seq")
-    @SequenceGenerator(name = "work_seq", sequenceName = "work_seq")
-    @Column(name = "work_id")
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    /**
-     * @return the publicationDate
-     */
-    @Column(name = "publication_date")
-    public PublicationDateEntity getPublicationDate() {
-        return publicationDate;
-    }
-
-    /**
-     * @param publicationDate
-     *            the publicationDate to set
-     */
-    public void setPublicationDate(PublicationDateEntity publicationDate) {
-        this.publicationDate = publicationDate;
-    }
-
-    /**
-     * @return the titles
-     */
-    @Column(name = "title", length = 1000)
-    public String getTitle() {
-        return title;
-    }
-
-    /**
-     * @param title
-     *            the title to set
-     */
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    @Column(name = "subtitle", length = 1000)
-    public String getSubtitle() {
-        return subtitle;
-    }
-
-    public void setSubtitle(String subtitle) {
-        this.subtitle = subtitle;
-    }
-
-    @Column(name = "description", length = 5000)
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    @Column(name = "work_url", length = 350)
-    public String getWorkUrl() {
-        return workUrl;
-    }
-
-    public void setWorkUrl(String workUrl) {
-        this.workUrl = workUrl;
-    }
 
     @Column(name = "citation", length = 5000)
     public String getCitation() {
@@ -152,43 +60,6 @@ public class WorkEntity extends BaseEntity<Long> implements Comparable<WorkEntit
         this.citation = citation;
     }
 
-    @Column(name = "journal_title", length = 1000)
-    public String getJournalTitle() {
-        return journalTitle;
-    }
-
-    public void setJournalTitle(String journalTitle) {
-        this.journalTitle = journalTitle;
-    }
-
-    
-    @Column(name = "language_code", length = 25)
-    public String getLanguageCode() {
-        return languageCode;
-    }
-
-    public void setLanguageCode(String languageCode) {
-        this.languageCode = languageCode;
-    }
-
-    @Column(name = "translated_title", length = 25)
-    public String getTranslatedTitle() {
-        return translatedTitle;
-    }
-
-    public void setTranslatedTitle(String translatedTitle) {
-        this.translatedTitle = translatedTitle;
-    }
-
-    @Column(name = "translated_title_language_code", length = 25)
-    public String getTranslatedTitleLanguageCode() {
-        return translatedTitleLanguageCode;
-    }
-
-    public void setTranslatedTitleLanguageCode(String translatedTitleLanguageCode) {
-        this.translatedTitleLanguageCode = translatedTitleLanguageCode;
-    }
-    
     @Basic
     @Enumerated(EnumType.STRING)
     @Column(name = "citation_type", length = 100)
@@ -200,17 +71,6 @@ public class WorkEntity extends BaseEntity<Long> implements Comparable<WorkEntit
         this.citationType = citationType;
     }
 
-    @Basic
-    @Enumerated(EnumType.STRING)
-    @Column(name = "work_type", length = 100)
-    public WorkType getWorkType() {
-        return workType;
-    }
-
-    public void setWorkType(WorkType workType) {
-        this.workType = workType;
-    }
-            
     @Column(name = "contributors_json")
     public String getContributorsJson() {
         return contributorsJson;
@@ -220,15 +80,6 @@ public class WorkEntity extends BaseEntity<Long> implements Comparable<WorkEntit
         this.contributorsJson = contributorsJson;
     }
 
-    @Column(name = "external_ids_json")
-    public String getExternalIdentifiersJson() {
-        return externalIdentifiersJson;
-    }
-
-    public void setExternalIdentifiersJson(String externalIdentifiersJson) {
-        this.externalIdentifiersJson = externalIdentifiersJson;
-    }    
-    
     @Basic
     @Enumerated(EnumType.STRING)
     @Column(name = "iso2_country", length = 2)
@@ -256,15 +107,7 @@ public class WorkEntity extends BaseEntity<Long> implements Comparable<WorkEntit
     public void setProfile(ProfileEntity profile) {
         this.profile = profile;
     }
-    
-    public SourceEntity getSource() {
-        return source;
-    }
 
-    public void setSource(SourceEntity source) {
-        this.source = source;
-    }
-    
     @Column(name = "added_to_profile_date")
     public Date getAddedToProfileDate() {
         return addedToProfileDate;
@@ -274,43 +117,16 @@ public class WorkEntity extends BaseEntity<Long> implements Comparable<WorkEntit
         this.addedToProfileDate = addedToProfileDate;
     }
 
-    @Basic
-    @Enumerated(EnumType.STRING)
-    public Visibility getVisibility() {
-        return visibility;
-    }
-
-    public void setVisibility(Visibility visibility) {
-        this.visibility = visibility;
-    }
-    
-    /*
-     * Dictates the display order for works (and versions of works)
-     * works with higher numbers should be displayed first. 
-     * 
-     * Currently only updatable via ProfileWorkDaoImpl.updateToMaxDisplay
-     *
-     */
-
-    @Column(name = "display_index", updatable=false, insertable=false)
-    public Long getDisplayIndex() {
-        return displayIndex;
-    }
-
-    public void setDisplayIndex(Long displayIndex) {
-        this.displayIndex = displayIndex;
-    }
-    
     @Override
     public int compareTo(WorkEntity other) {
         if (other == null) {
             throw new NullPointerException("Can't compare with null");
         }
 
-        int comparison = compareOrcidId(other);        
+        int comparison = compareOrcidId(other);
         if (comparison == 0) {
             comparison = comparePublicationDate(other);
-            if(comparison == 0) {
+            if (comparison == 0) {
                 comparison = compareTitles(other);
                 if (comparison == 0) {
                     return compareIds(other);
@@ -370,19 +186,19 @@ public class WorkEntity extends BaseEntity<Long> implements Comparable<WorkEntit
     }
 
     private int compareOrcidId(WorkEntity other) {
-        if(this.getProfile() == null) {
-            if(other.getProfile() == null) {
+        if (this.getProfile() == null) {
+            if (other.getProfile() == null) {
                 return 0;
             } else {
                 return -1;
             }
-        } else if(other.getProfile() == null) {
+        } else if (other.getProfile() == null) {
             return 1;
         } else {
             return this.getProfile().getId().compareTo(other.getProfile().getId());
-        }        
+        }
     }
-    
+
     public static class ChronologicallyOrderedWorkEntityComparator implements Comparator<WorkEntity> {
         public int compare(WorkEntity work1, WorkEntity work2) {
             if (work2 == null) {
