+++ orcid-persistence/src/main/resources/META-INF/persistence.xml
@@ -32,6 +32,7 @@
         <class>org.orcid.persistence.jpa.entities.FuzzyDateEntity</class>
         <class>org.orcid.persistence.jpa.entities.GivenPermissionByEntity</class>
         <class>org.orcid.persistence.jpa.entities.GivenPermissionToEntity</class>
+        <class>org.orcid.persistence.jpa.entities.MinimizedWorkEntity</class>
         <class>org.orcid.persistence.jpa.entities.NotificationItemEntity</class>
         <class>org.orcid.persistence.jpa.entities.NotificationAddItemsEntity</class>
         <class>org.orcid.persistence.jpa.entities.NotificationAmendedEntity</class>
@@ -56,7 +57,7 @@
         <class>org.orcid.persistence.jpa.entities.ShibbolethAccountEntity</class>
         <class>org.orcid.persistence.jpa.entities.SubjectEntity</class>
         <class>org.orcid.persistence.jpa.entities.WebhookEntity</class>
-        <class>org.orcid.persistence.jpa.entities.WorkEntity</class>        
+        <class>org.orcid.persistence.jpa.entities.WorkEntity</class>
         <class>org.orcid.persistence.jpa.entities.GroupIdRecordEntity</class>
         <class>org.orcid.persistence.jpa.entities.UserconnectionEntity</class>
 		<class>org.orcid.persistence.jpa.entities.UserconnectionPK</class>
